+++ src/redpanda_http_fake.py
@@ -181,7 +181,7 @@ if __name__ == "__main__":
             epoch = now + pendulum.duration(microseconds=idx * (1000000 / args.rate))
 
             value = {
-                "timestamp": epoch.timestamp() * 1e6,
+                "timestamp": int(epoch.timestamp() * 1e6),
                 **key_vals,
                 **fake.values(),
             }
